Menu Features

+++ Item.java
@@ -5,7 +5,7 @@
 public class Item {
     private String name;
     private float price;
-    private int calories;
+    private float calories;
     private int stock;
     static int itemCount = 0;
 
@@ -17,7 +17,7 @@ public class Item {
      * @param calories      the calories of the item
      * @param initStock     the initial stock quantity of the item
      */
-    public Item(String name, float price, int calories, int initStock) {
+    public Item(String name, float price, float calories, int initStock) {
         this.name = name;
         this.price = price;
         this.calories = calories;
@@ -56,7 +56,7 @@ public class Item {
      *
      * @return the calories of the item
      */
-    public int getCalories() {
+    public float getCalories() {
         return calories;
     }
